refactor(drive-integration): moving thigs from wizardSteps.ts to AddEntryWizard.tsx
